Lay out package structure & add log analysis tool
This change adds a package skeleton for E3Analysis and
a simple command-line tool for parsing raw udp2log log lines
generated by the ClickTracking extension. I know there are
other existing solutions -- I'm not attached to this one so
if it isn't useful for you I'm happy to abandon it. Feedback
appreciated.

Change-Id: Iefa4dec1040bd0ee7ba6d07556fb3838b1ab039f